Revert "Added font scaling"
This reverts commit c9b01958beec072c2b42d0c5773bb0035bec3fb4.

+++ GameEngine/TextRenderer.h
@@ -12,7 +12,7 @@
 
 struct Character
 {
-	unsigned int textureAtlasOffset;
+	float textureAtlasOffset;
 	glm::ivec2 size;
 	glm::ivec2 bearing;
 	glm::ivec2 advance;
